+++ tests-e2e/cypress/integration/playbooks/edit_spec.js
@@ -198,13 +198,13 @@ describe('playbooks > edit', () => {
 
                     cy.get('#create-channel').within(() => {
                         // * Verify that the toggle is checked
-                        cy.get('label input').should('be.checked');
+                        cy.get('[type="checkbox"]').should('be.checked');
 
                         // # Click on the toggle to enable the setting
-                        cy.get('label input').click({force: true});
+                        cy.get('[type="checkbox"]').click({force: true});
 
                         // * Verify that the toggle remains checked
-                        cy.get('label input').should('be.checked');
+                        cy.get('[type="checkbox"]').should('be.checked');
                     });
                 });
             });
@@ -820,8 +820,8 @@ describe('playbooks > edit', () => {
                     // # Visit the selected playbook
                     cy.visit(`/playbooks/playbooks/${testPlaybook.id}/edit`);
 
-                    // # Switch to Actions tab
-                    cy.get('#root').findByText('Actions').click();
+                    // # Switch to Status updates tab
+                    cy.get('#root').findByText('Status updates').click();
 
                     // * Verify that the toggle is unchecked
                     cy.get('#broadcast-channels label input').should(
@@ -834,7 +834,7 @@ describe('playbooks > edit', () => {
                     cy.visit(`/playbooks/playbooks/${testPlaybook.id}/edit`);
 
                     // # Switch to Actions tab
-                    cy.get('#root').findByText('Actions').click();
+                    cy.get('#root').findByText('Status updates').click();
 
                     cy.get('#broadcast-channels').within(() => {
                         // * Verify that the toggle is unchecked
@@ -852,8 +852,8 @@ describe('playbooks > edit', () => {
                     // # Visit the selected playbook
                     cy.visit(`/playbooks/playbooks/${testPlaybook.id}/edit`);
 
-                    // # Switch to Actions tab
-                    cy.get('#root').findByText('Actions').click();
+                    // # Switch to Status updates tab
+                    cy.get('#root').findByText('Status updates').click();
 
                     // * Verify that the toggle is unchecked
                     cy.get('#broadcast-channels label input').should(
@@ -873,8 +873,8 @@ describe('playbooks > edit', () => {
                     // # Visit the selected playbook
                     cy.visit(`/playbooks/playbooks/${testPlaybook.id}/edit`);
 
-                    // # Switch to Actions tab
-                    cy.get('#root').findByText('Actions').click();
+                    // # Switch to Status updates tab
+                    cy.get('#root').findByText('Status updates').click();
 
                     cy.get('#broadcast-channels').within(() => {
                         // * Verify that the toggle is unchecked
@@ -903,8 +903,8 @@ describe('playbooks > edit', () => {
                     // # Visit the selected playbook
                     cy.visit(`/playbooks/playbooks/${testPlaybook.id}/edit`);
 
-                    // # Switch to Actions tab
-                    cy.get('#root').findByText('Actions').click();
+                    // # Switch to Status updates tab
+                    cy.get('#root').findByText('Status updates').click();
 
                     cy.get('#broadcast-channels').within(() => {
                         // * Verify that the toggle is unchecked
@@ -946,8 +946,8 @@ describe('playbooks > edit', () => {
                     // # Visit the selected playbook
                     cy.visit(`/playbooks/playbooks/${testPlaybook.id}/edit`);
 
-                    // # Switch to Actions tab
-                    cy.get('#root').findByText('Actions').click();
+                    // # Switch to Status updates tab
+                    cy.get('#root').findByText('Status updates').click();
 
                     cy.get('#broadcast-channels').within(() => {
                         // * Verify that the toggle is unchecked
@@ -984,7 +984,7 @@ describe('playbooks > edit', () => {
                     cy.visit(`/playbooks/playbooks/${testPlaybook.id}/edit`);
 
                     // # Switch to Actions tab
-                    cy.get('#root').findByText('Actions').click();
+                    cy.get('#root').findByText('Status updates').click();
 
                     cy.get('#broadcast-channels').within(() => {
                         // * Verify that the toggle is unchecked
@@ -1042,8 +1042,8 @@ describe('playbooks > edit', () => {
                             // # Navigate again to the playbook
                             cy.visit(`/playbooks/playbooks/${playbookId}/edit`);
 
-                            // # Switch to Actions tab
-                            cy.get('#root').findByText('Actions').click();
+                            // # Switch to Status updates tab
+                            cy.get('#root').findByText('Status updates').click();
 
                             // # Save the playbook
                             cy.findByTestId('save_playbook').click();
@@ -1057,8 +1057,8 @@ describe('playbooks > edit', () => {
                             // # Navigate again to the playbook
                             cy.visit(`/playbooks/playbooks/${playbookId}/edit`);
 
-                            // # Switch to Actions tab
-                            cy.get('#root').findByText('Actions').click();
+                            // # Switch to Status updates tab
+                            cy.get('#root').findByText('Status updates').click();
 
                             cy.get('#broadcast-channels').within(() => {
                                 // * Verify that the toggle is unchecked
@@ -1081,8 +1081,8 @@ describe('playbooks > edit', () => {
                     // # Visit the selected playbook
                     cy.visit('/playbooks/playbooks/' + testPlaybook.id + '/edit');
 
-                    // # Switch to Templates tab
-                    cy.get('#root').findByText('Actions').click();
+                    // # Switch to Status updates tab
+                    cy.get('#root').findByText('Status updates').click();
 
                     // * Verify placeholder text is present
                     cy.get('#playbook-automation-broadcast').should(
@@ -1095,8 +1095,8 @@ describe('playbooks > edit', () => {
                     // # Visit the selected playbook
                     cy.visit('/playbooks/playbooks/' + testPlaybook.id + '/edit');
 
-                    // # Switch to Actions tab
-                    cy.get('#root').findByText('Actions').click();
+                    // # Switch to Status updates tab
+                    cy.get('#root').findByText('Status updates').click();
 
                     // # Click on the toggle to enable the setting
                     cy.get('#broadcast-channels').within(() => {
@@ -1114,8 +1114,8 @@ describe('playbooks > edit', () => {
                     // # Visit the selected playbook
                     cy.visit('/playbooks/playbooks/' + testPlaybook.id + '/edit');
 
-                    // # Switch to Actions tab
-                    cy.get('#root').findByText('Actions').click();
+                    // # Switch to Status updates tab
+                    cy.get('#root').findByText('Status updates').click();
 
                     // * Verify placeholder text is present
                     cy.get('#playbook-automation-broadcast').should(
@@ -1128,8 +1128,8 @@ describe('playbooks > edit', () => {
                     // # Visit the selected playbook
                     cy.visit('/playbooks/playbooks/' + testPlaybook.id + '/edit');
 
-                    // # Switch to Actions tab
-                    cy.get('#root').findByText('Actions').click();
+                    // # Switch to Status updates tab
+                    cy.get('#root').findByText('Status updates').click();
 
                     // # Click on the toggle to enable the setting
                     cy.get('#broadcast-channels').within(() => {
@@ -1147,8 +1147,8 @@ describe('playbooks > edit', () => {
                     // # Visit the selected playbook
                     cy.visit('/playbooks/playbooks/' + testPlaybook.id + '/edit');
 
-                    // # Switch to Actions tab
-                    cy.get('#root').findByText('Actions').click();
+                    // # Switch to Status updates tab
+                    cy.get('#root').findByText('Status updates').click();
 
                     // * Verify placeholder text is present
                     cy.get('#playbook-automation-broadcast').should(
@@ -1161,8 +1161,8 @@ describe('playbooks > edit', () => {
                     // # Visit the selected playbook
                     cy.visit('/playbooks/playbooks/' + testPlaybook.id + '/edit');
 
-                    // # Switch to Actions tab
-                    cy.get('#root').findByText('Actions').click();
+                    // # Switch to Status updates tab
+                    cy.get('#root').findByText('Status updates').click();
 
                     // # Click on the toggle to enable the setting
                     cy.get('#broadcast-channels').within(() => {
@@ -1187,8 +1187,8 @@ describe('playbooks > edit', () => {
                     // # Visit the selected playbook
                     cy.visit('/playbooks/playbooks/' + testPlaybook.id + '/edit');
 
-                    // # Switch to Actions tab
-                    cy.get('#root').findByText('Actions').click();
+                    // # Switch to Status updates tab
+                    cy.get('#root').findByText('Status updates').click();
 
                     // * Verify placeholder text is present
                     cy.get('#playbook-automation-broadcast').should(
@@ -1375,7 +1375,7 @@ describe('playbooks > edit', () => {
                     // # Type name to use new custom category
                     cy.get('#playbook-automation-categorize-playbook-run')
                         .click()
-                        .type('Custom category' + '{enter}', {delay: 200});
+                        .type('Custom category{enter}', {delay: 200});
 
                     // # Save the playbook
                     cy.findByTestId('save_playbook').click();
@@ -1399,13 +1399,14 @@ describe('playbooks > edit', () => {
                 });
             });
         });
+
         describe('status updates enable / disabled', () => {
             it('is enabled in a new playbook', () => {
                 // # Visit the selected playbook
                 cy.visit(`/playbooks/playbooks/${testPlaybook.id}/edit`);
 
-                // # Switch to Templates tab
-                cy.get('#root').findByText('Templates').click();
+                // # Switch to Status updates tab
+                cy.get('#root').findByText('Status updates').click();
 
                 // * Verify that the toggle is checked
                 cy.get('#status-updates label input').should('be.checked');
@@ -1415,8 +1416,8 @@ describe('playbooks > edit', () => {
                 // # Visit the selected playbook
                 cy.visit(`/playbooks/playbooks/${testPlaybook.id}/edit`);
 
-                // # Switch to Templates tab
-                cy.get('#root').findByText('Templates').click();
+                // # Switch to Status updates tab
+                cy.get('#root').findByText('Status updates').click();
 
                 // * Verify that toggle can be disabled
                 cy.get('#status-updates').within(() => {
@@ -1440,9 +1441,6 @@ describe('playbooks > edit', () => {
 
                 // * Verify that the update text is disabled
                 cy.get('#playbook_reminder_message_edit').should('be.disabled');
-
-                // # Switch to Actions tab
-                cy.get('#root').findByText('Actions').click();
 
                 // * Verify that the toggle can't be checked
                 cy.get('#broadcast-channels').within(() => {
@@ -1473,8 +1471,8 @@ describe('playbooks > edit', () => {
                 // # Visit the selected playbook
                 cy.visit(`/playbooks/playbooks/${testPlaybook.id}/edit`);
 
-                // # Switch to Actions tab
-                cy.get('#root').findByText('Actions').click();
+                // # Switch to Status updates tab
+                cy.get('#root').findByText('Status updates').click();
 
                 // * Verify that the toggle is checked
                 cy.get('#broadcast-channels').within(() => {
@@ -1500,9 +1498,6 @@ describe('playbooks > edit', () => {
                     cy.get('label input').should('be.checked');
                 });
 
-                // # Switch to Templates tab
-                cy.get('#root').findByText('Templates').click();
-
                 // * Verify that toggle can be selected
                 cy.get('#status-updates').within(() => {
                     // * Verify that the toggle is checked
@@ -1514,9 +1509,6 @@ describe('playbooks > edit', () => {
                     // * Verify that the toggle is unchecked
                     cy.get('label input').should('not.be.checked');
                 });
-
-                // # Switch to Actions tab
-                cy.get('#root').findByText('Actions').click();
 
                 // * Verify that the toggle can't be checked
                 cy.get('#broadcast-channels').within(() => {
@@ -1547,8 +1539,8 @@ describe('playbooks > edit', () => {
                 // # Visit the selected playbook
                 cy.visit(`/playbooks/playbooks/${testPlaybook.id}/edit`);
 
-                // # Switch to Templates tab
-                cy.get('#root').findByText('Templates').click();
+                // # Switch to Status updates tab
+                cy.get('#root').findByText('Status updates').click();
 
                 // # Uncheck toggle
                 cy.get('#status-updates label input').click({force: true});
@@ -1559,8 +1551,8 @@ describe('playbooks > edit', () => {
                 // # Navigate again to the playbook
                 cy.visit(`/playbooks/playbooks/${testPlaybook.id}/edit`);
 
-                // # Switch to Templates tab
-                cy.get('#root').findByText('Templates').click();
+                // # Switch to Status updates tab
+                cy.get('#root').findByText('Status updates').click();
 
                 // * Verify that the toggle is unchecked
                 cy.get('#status-updates label input').should('not.be.checked');
@@ -1572,8 +1564,8 @@ describe('playbooks > edit', () => {
                 // # Visit the selected playbook
                 cy.visit(`/playbooks/playbooks/${testPlaybook.id}/edit`);
 
-                // # Switch to Templates tab
-                cy.get('#root').findByText('Templates').click();
+                // # Switch to Retrospective tab
+                cy.get('#root').findByText('Retrospective').click();
 
                 // * Verify that the toggle is checked
                 cy.get('#retrospective-enabled label input').should('be.checked');
@@ -1583,8 +1575,8 @@ describe('playbooks > edit', () => {
                 // # Visit the selected playbook
                 cy.visit(`/playbooks/playbooks/${testPlaybook.id}/edit`);
 
-                // # Switch to Templates tab
-                cy.get('#root').findByText('Templates').click();
+                // # Switch to Retrospective tab
+                cy.get('#root').findByText('Retrospective').click();
 
                 // * Verify that toggle can be disabled
                 cy.get('#retrospective-enabled').within(() => {
@@ -1614,8 +1606,8 @@ describe('playbooks > edit', () => {
                 // # Visit the selected playbook
                 cy.visit(`/playbooks/playbooks/${testPlaybook.id}/edit`);
 
-                // # Switch to Templates tab
-                cy.get('#root').findByText('Templates').click();
+                // # Switch to Retrospective tab
+                cy.get('#root').findByText('Retrospective').click();
 
                 // # Uncheck toggle
                 cy.get('#retrospective-enabled label input').click({force: true});
@@ -1626,8 +1618,8 @@ describe('playbooks > edit', () => {
                 // # Navigate again to the playbook
                 cy.visit(`/playbooks/playbooks/${testPlaybook.id}/edit`);
 
-                // # Switch to Templates tab
-                cy.get('#root').findByText('Templates').click();
+                // # Switch to Retrospective tab
+                cy.get('#root').findByText('Retrospective').click();
 
                 // * Verify that the toggle is unchecked
                 cy.get('#retrospective-enabled label input').should('not.be.checked');
