+++ webapp/i18n/en.json
@@ -2,7 +2,6 @@
   "+8G9qr": "Default text for the retrospective.",
   "+QgvjN": "Assign the owner role to",
   "+Tmpup": "You automatically receive updates when this playbook is run.",
-  "+ZIXOR": "Channel access",
   "+hddg7": "Add to run timeline",
   "/1FEJW": "ACTIVE PARTICIPANTS per day over the last 14 days",
   "/4tOwT": "Skip",
@@ -16,6 +15,7 @@
   "0Vvpht": "Make Playbook Member",
   "0oL1zz": "Copied!",
   "0oLj/t": "Expand",
+  "0q+hj2": "Define a template for a concise description that explains each run to its stakeholders.",
   "0tznw6": "Convert to private playbook",
   "0wJ7N+": "Task",
   "15jbT0": "Add more to your timeline",
@@ -43,7 +43,6 @@
   "5CI3KH": "Contact support",
   "5FRgqE": "Downloading channel log",
   "5Ofkag": "Enable retrospective",
-  "5Ot7cd": "Determine the type of channel this playbook creates.",
   "5ciuDD": "NOT IN CHANNEL",
   "5j6GD/": "{numParticipants, plural, =0 {no participants} =1 {# participant} other {# participants}}",
   "5qBEKB": "What are playbook runs?",
@@ -65,7 +64,6 @@
   "9tBhzB": "Upgrade now",
   "9uOFF3": "Overview",
   "A21Mgv": "Run finished",
-  "A3ptul": "Templates",
   "A8dbCS": "Playbook Not Found",
   "AF9wda": "This update will be saved to the <OverviewLink>overview page</OverviewLink>{hasBroadcast, select, true { and broadcast to <ChannelsTooltip>{broadcastChannelCount, plural, =1 {one channel} other {{broadcastChannelCount, number} channels}}</ChannelsTooltip>} other {}}.",
   "AML4RW": "Task assignments",
@@ -175,7 +173,6 @@
   "S0kWcH": "Update overdue",
   "SDSqfA": "When a run starts",
   "SENRqu": "Help",
-  "SFuk1v": "Permissions",
   "SXJ98n": "You will not be able to edit the retrospective report after publishing it. Do you want to publish the retrospective report?",
   "SmAUf9": "A reminder will be sent <b>{timestamp}</b>",
   "T5rX+W": "How often should an update be posted?",
@@ -303,6 +300,7 @@
   "q0cpUe": "Add checklist",
   "q6f8x9": "Change since last update",
   "qp3Fk4": "A playbook is a workflow that your teams and tools should follow, including everything from checklists, actions, templates, and retrospectives.",
+  "qsr3Zk": "Update the run Summary",
   "qyJtWy": "Show less",
   "rDvvQs": "{completed, number} / {total, number} done",
   "rX08cW": "Date must be in the future.",
@@ -325,7 +323,6 @@
   "tzMNF3": "Status",
   "u4MwUB": "Save your playbook run history",
   "uBLF+D": "What is a playbook?",
-  "uJ3bRR": "This template helps to standardize the format for a concise description that explains each run to its stakeholders.",
   "uhu5aG": "Public",
   "uny3Zy": "Playbooks",
   "usa8vQ": "Send a welcome message",
