+++ webapp/src/components/backstage/playbook_edit/automation/clear_indicator.tsx
