+++ webapp/src/components/backstage/playbook_edit/automation/input_keywords.tsx
@@ -8,8 +8,8 @@ import {OptionTypeBase, StylesConfig} from 'react-select';
 
 import CreatableSelect from 'react-select/creatable';
 
-import {AutomationHeader, AutomationTitle, SelectorWrapper} from 'src/components/backstage/automation/styles';
-import {Toggle} from 'src/components/backstage/automation/toggle';
+import {AutomationHeader, AutomationTitle, SelectorWrapper} from 'src/components/backstage/playbook_edit/automation/styles';
+import {Toggle} from 'src/components/backstage/playbook_edit/automation/toggle';
 
 interface Props {
     enabled: boolean;
