+++ webapp/src/components/backstage/playbook_edit/styles.tsx
@@ -1,0 +1,35 @@
+// Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
+// See LICENSE.txt for license information.
+
+import styled from 'styled-components';
+
+export const Section = styled.div`
+    margin: 32px 0;
+`;
+
+export const SectionTitle = styled.div`
+    font-weight: 600;
+    margin: 0 0 32px 0;
+`;
+
+export const SidebarBlock = styled.div`
+    margin: 0 0 40px;
+`;
+
+export const Setting = styled.div`
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+
+    margin-bottom: 24px;
+`;
+
+export const BackstageGroupToggleHeader = styled.div`
+    font-weight: 600;
+    font-size: 16px;
+    line-height: 24px;
+    color: var(--center-channel-color);
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+`;
